pouw_dem/training/train_real_agents.py
#!/usr/bin/env python3
"""
REAL Training Script for FDRL Agents
This uses the unified environment and DSAC agents
"""

import sys
import os
sys.path.append(os.path.dirname(os.path.abspath(__file__)))

from create_unified_training import UnifiedPoUWEnvironment, create_dsac_agent
import numpy as np
import torch
from datetime import datetime

def train_agents(n_episodes=100):
    """Train FDRL agents on unified environment"""
    
    print("🚀 STARTING REAL FDRL TRAINING")
    print("="*70)
    
    # Create environments and agents for each facility
    facilities = [
        {'id': 0, 'name': 'MARA_TX_1', 'capacity': 100},
        {'id': 1, 'name': 'MARA_TX_2', 'capacity': 80},
        {'id': 2, 'name': 'MARA_MT_1', 'capacity': 60}
    ]
    
    envs = []
    agents = []
    
    for facility in facilities:
        env = UnifiedPoUWEnvironment(
            house_id=facility['id'],
            capacity_mw=facility['capacity']
        )
        envs.append(env)
        
        agent = create_dsac_agent(env, facility['id'], facility['capacity'])
        agents.append(agent)
    
    # Training loop
    for episode in range(n_episodes):
        states = [env.reset() for env in envs]
        episode_rewards = [0] * len(agents)
        
        done = False
        step = 0
        
        while not done:
            # Each agent selects action
            actions = []
            for i, (agent, state) in enumerate(zip(agents, states)):
                action = agent.get_action(state)
                actions.append(action)
            
            # Execute actions
            next_states = []
            for i, (env, action) in enumerate(zip(envs, actions)):
                next_state, reward, done, info = env.step(action)
                next_states.append(next_state)
                episode_rewards[i] += reward
                
                # Train agent (if implemented)
                # agent.train_step(state, action, reward, next_state, done)
            
            states = next_states
            step += 1
        
        # Print progress
        if (episode + 1) % 10 == 0:
            print(f"\nEpisode {episode + 1}/{n_episodes}")
            for i, facility in enumerate(facilities):
                print(f"  {facility['name']}: Total Reward = {episode_rewards[i]:.2f}")
    
    print("\n✅ Training Complete!")
    return agents

if __name__ == "__main__":
    trained_agents = train_agents(n_episodes=10)  # Quick test
